+++ libearth/schema.py
@@ -1731,18 +1731,23 @@ def validate(element, recurse=True, raise_error=True):
                 )
             return False
     for name, desc in inspect_child_tags(element_type).values():
-        child_element = getattr(element, name, None)
-        if desc.required and not child_element:
-            if raise_error:
-                raise IntegrityError(
-                    '{0.__module__}.{0.__name__}.{1} is required, but '
-                    '{2!r} lacks it'.format(element_type, name, element)
-                )
-            return False
-        if recurse and child_element is not None:
-            if validate(child_element, recurse=True, raise_error=raise_error):
-                continue
-            return False
+        children = getattr(element, name, None)
+        if not desc.multiple:
+            children = children,
+        for child_element in children:
+            if desc.required and not child_element:
+                if raise_error:
+                    raise IntegrityError(
+                        '{0.__module__}.{0.__name__}.{1} is required, but '
+                        '{2!r} lacks it'.format(element_type, name, element)
+                    )
+                return False
+            if recurse and child_element is not None:
+                if validate(child_element,
+                            recurse=True,
+                            raise_error=raise_error):
+                    continue
+                return False
     return True
 
 
@@ -1860,7 +1865,10 @@ class write(collections.Iterable):
                 yield ':'
             yield desc.name
             yield '='
-            yield encode(quoteattr(encoded_attr_value))
+            quoted_attr = quoteattr(encoded_attr_value)
+            if not isinstance(quoteattr, binary_type):
+                quoted_attr = encode(quoted_attr)
+            yield quoted_attr
         content = inspect_content_tag(element_type)
         children = inspect_child_tags(element_type)
         escape = xml.sax.saxutils.escape