+++ docs/changes.rst
@@ -8,6 +8,8 @@ To be released.
 
 - Added a workaround for thread unsafety :func:`time.strftime()` on CPython.
   See http://bugs.python.org/issue7980 as well.  [:issue:`32`]
+- Fixed :exc:`UnicodeDecodeError` which is raised when a feed title contains
+  any non-ASCII characters.  [:issue:`34` by Jae-Myoung Yu]
 
 
 Version 0.1.0