+++ docs/changes.rst
@@ -4,7 +4,7 @@ Libearth Changelog
 Version 0.1.1
 -------------
 
-To be released.
+Released on January 2, 2014.
 
 - Added a workaround for thread unsafety :func:`time.strftime()` on CPython.
   See http://bugs.python.org/issue7980 as well.  [:issue:`32`]